### 2017-04-16 #241 @labarba
The code repository needs *A statement of need*, which means ...

> The authors should clearly state what problems the software is designed to solve and who the target audience is.

I was not able to find this, @bjmorgan.
