### 2017-04-17 845b77e
More information about the scope of the problem targetted by the software added to the README and paper.md
