+++ paper.bib
@@ -59,3 +59,52 @@
 	Volume = {36},
 	Year = {1977}}
 
+@inbook{Voter_RadiationEffectsInSolids2007,
+	Address = {Dordrecht},
+	Author = {Voter, Arthur F.},
+	Booktitle = {Radiation Effects in Solids},
+	Date-Added = {2017-04-14 12:20:57 +0000},
+	Date-Modified = {2017-04-14 12:21:24 +0000},
+	Doi = {10.1007/978-1-4020-5295-8_1},
+	Editor = {Sickafus, Kurt E. and Kotomin, Eugene A. and Uberuaga, Blas P.},
+	Isbn = {978-1-4020-5295-8},
+	Pages = {1--23},
+	Publisher = {Springer Netherlands},
+	Title = {Introduction to the Kinetic Monte Carlo Method},
+	Url = {http://dx.doi.org/10.1007/978-1-4020-5295-8_1},
+	Year = {2007},
+	Bdsk-Url-1 = {http://dx.doi.org/10.1007/978-1-4020-5295-8_1}}
+
+@article{Catlow_SolidStateIonics1983,
+	Author = {Catlow, C. Richard A.},
+	Date-Added = {2012-12-20 17:14:00 +0000},
+	Date-Modified = {2012-12-20 17:14:37 +0000},
+	Journal = {Sol. Stat. Ionics},
+	Pages = {89--107},
+	Title = {Static Lattice Simulation of Structure and Transport in Superionic Conductors},
+	Volume = {8},
+	Year = {1983}}
+
+
+@article{HowardAndLidiard_RepProgPhys1964,
+	Author = {Howard, R. E. and Lidiard, A. B.},
+	Date-Added = {2013-06-17 10:49:20 +0000},
+	Date-Modified = {2013-06-17 10:49:20 +0000},
+	Journal = {Rep Prog Phys.},
+	Pages = {161--240},
+	Title = {Matter Transport in Solids},
+	Volume = {27},
+	Year = {1964}}
+
+@inbook{Stoneham_IonicSolidsAtHighTemperatures,
+	Author = {Harding, J. F.},
+        Title = {Defects and Transport in Ionic Solids},
+        Date-Added = {2016-12-09 16:41:41 +0000},
+	Date-Modified = {2016-12-09 16:43:25 +0000},
+	Editor = {Stoneham, A. M.},
+	Publisher = {World Scientific},
+	Booktitle = {Ionic Solids at High Temperatures},
+	Year = {1989},
+        Doi = {10.1142/9789814503228_0003},
+        Url = {https://doi.org/10.1142/9789814503228_0003}}
+
