+++ paper.md
@@ -17,9 +17,13 @@ bibliography: paper.bib
 
 # Summary
 
-`lattice_mc` is Python module for running (kinetic) lattice-gas Monte Carlo simulations. 
-Simple lattices can be constructed programmatically (presently square, honeycomb, and cubic lattices).
-Arbitrary lattices can be generated from files that define the lattice sites and their connectivity. The algorithms used and interaction models are described in <a href="#ref1">\[1\]</a>. Calculated properties include tracer and &ldquo;jump&rdquo; diffusion coefficients; where the latter is proportional to the mobility (and hence the conductivity for charged particles) <a href="#ref2">\[2\]</a>; and tracer (single particle) and collective correlation factors, f and f_I <a href="#ref3">\[3\]</a>. The simplest interaction model is for &ldquo;non-interacting&rdquo; particles, where the only restriction is volume exclusion (two particles cannot simultaneoulsy occupy a single site) <a href="#ref1">\[4\]</a>. Additional interaction models include nearest-neighbour repulsion and on-site energies for inequivalent sites.
+`lattice_mc` is Python module for performing (kinetic) lattice-gas Monte Carlo (LGMC) simulations of ionic transport in solid electrolytes.
+
+In solid electrolytes, ionic motion is typically effected by a series of discrete "jumps" where ions move between adjacent lattice sites \[[1](#Catlow_SolStatIonics1983)\]. For dilute mobile ions, ionic trajectories are random walks, and simple analytical expressions exits that relate macroscopic transport coefficients, i.e. diffusion coefficients and ionic conductivities, to the microscopic jump frequency of individual ions \[[2](#HowardAndLidiard_RepProgPhys1964),[3](#Stoneham_IonicSolids1989)\]. Practical solid electrolytes have high mobile ion concentrations, with significnat interparticle interactions producing deviations from the dilute limit random walk behaviour. In general, the *quantitative* effect of interparticle interactions cannot be determined analytically. As an alternative, numerical simulations, such as lattice-gas Monte Carlo methods, can be used to directly calculate these relationships. Lattice-gas Monte Carlo methods are particularly suited to studying how varying properties across broad classes of materials give quantitative differences in macroscopic ionic transport, and can be used to understand the different transport properties of materials with, for example, different crystal structures or mobile ion stoichiometries. 
+
+`lattice_mc` has been written to allow materials scientists and solid-state chemists model how the microscopic physics of solid electrolytes (crystal structure, stoichiometry, interaction models) determine macroscopic transport behaviour (diffusion and ionic conduction), with the goal of understand the factors that make different materials more or less useful for specific applications (e.g. Li-ion batteries or fuel cells).
+
[…]
