### 2017-04-17 #241 @bjmorgan
@labarba I have amended the README.md and paper.md. Hopefully the additional information is suitable and appropriate for the JOSS requirements.
