### 2017-04-18 #241 @labarba
The title in the paper metadata appears as *"lattice_mc: A Python Lattice-Gas Monte Carlo Module"*. I"m not sure why this full title does not appear in the title of the pre-review issue ... it should also probably match the heading in the README.

I'm looking for a reviewer now.
