### 2017-04-20 #241 @labarba
@arfon : Does our bot have a command to _unassign_ a reviewer? I assigned myself by mistake.
