### 2017-04-20 #241 @arfon
> @arfon : Does our bot have a command to unassign a reviewer? I assigned myself by mistake.

Not right now sorry, you're on the hook for reviewing now :trollface:. In all seriousness, when we find a reviewer just use the `@whedon assign username as reviewer` and it will clobber your assignment.
