### 2017-04-22 #247 @whedon
[REVIEW]: lattice_mc: A Python Lattice-Gas Monte Carlo Module
**Submitting author:** @bjmorgan (<a href="http://orcid.org/0000-0002-3056-8233">Benjamin J. Morgan</a>)
**Repository:** <a href="https://github.com/bjmorgan/lattice_mc" target ="_blank">https://github.com/bjmorgan/lattice_mc</a>
**Version:** 1.0.0
