### 2017-05-01 #247 @asmit3
I was able to tick off all the boxes except the one on Community Guidelines since the LICENSE file appears incomplete (no year or full name of author). Also I didn't see an email address for correspondence.
The default unittest given in the README file throws 3 errors when I run on my mac with python 3.5 (see attachment)
![screenshot 2017-05-01 12 06 07](https://cloud.githubusercontent.com/assets/10836284/25590814/cb3d5f9c-2e66-11e7-9e77-cd67ebe677d4.png)

The ipython notebook was very useful in explaining how the code works.
