### 2017-05-02 695e679
3.6+ compatible tests rewritten for 3.5.0 compatibility Mock methods such as Mock.assert_called_once() are new in 3.6.0. For compatibility with 3.5+ these have been replaced with e.g. assertEqual( Mock.call_count, 1 )
