### 2017-05-02 eb706fb
Year and name added to LICENSE
