### 2017-05-02 0d3287b
email address added, and better spacing for older markdown processors
