### 2017-05-02 #247 @bjmorgan
The errors under python 3.5 were due to test methods that were introduced with version 3.6. I have replaced these with 3.5+ compatible versions:
https://github.com/bjmorgan/lattice_mc/commit/695e679e66e2331b6e6c23c70e983addf555ec38

I have added the year and my name to the LICENSE file:
https://github.com/bjmorgan/lattice_mc/commit/eb706fb5160b018ec134a532c49b49bd189c26b2

My email is given in `setup.py`, and I have added it to `paper.md`:
https://github.com/bjmorgan/lattice_mc/commit/0d3287bc5ce02823b032d29105998b6cd63a4195

I don't know if you would recommend including my email address elsewhere in the package?
