### 2017-05-02 #247 @asmit3
The changes are satisfactory for me. I don't have any more comments. Thanks
