### 2017-05-02 18fd825
Experimental travis config file added
