### 2017-05-02 596f021
Added scipy to list of requirements
