### 2017-05-02 2decc6a
Added Travis CI badge to README
