### 2017-05-02 7b2cb85
Removed 3.7 from Travis config
