+++ .travis.yml
@@ -11,4 +11,7 @@ python:
 # command to install dependencies
 install: "pip install -r requirements.txt"
 # command to run tests
-script: python -m unittest discover
+script: coverage run -m unittest discover
+after_success:
+coveralls -d tests/.coveralls
+
