### 2017-05-05 6ad582f
adjusted .travis.yml
