+++ .travis.yml
@@ -12,6 +12,5 @@ python:
 install: "pip install -r requirements.txt"
 # command to run tests
 script: coverage run -m unittest discover
-after_success:
-coveralls -d tests/.coveralls
+after_success: coveralls -d tests/.coveralls
 
