### 2017-05-05 9eaea67
Added coverage and python-coveralls to requirements
