+++ .travis.yml
@@ -9,8 +9,8 @@ python:
   - "3.6"
   - "3.6-dev" # 3.6 development branch
 # command to install dependencies
-install: "pip install -r requirements.txt"
+install: pip install -r requirements.txt
 # command to run tests
 script: coverage run -m unittest discover
-after_success: coveralls -d tests/.coveralls
+after_success: coveralls
 
