### 2017-05-05 1360789
fiddling with coveralls settings
