### 2017-05-05 0587ba2
Removed email notifications from Travis CI builds
