### 2017-05-05 d653021
Added test coverage badge to README
