### 2017-05-05 5a0e45c
Code Climate test reporting integration
