### 2017-05-05 6737260
Trying Code Climate for test coverage reporting
