+++ requirements.txt
@@ -3,5 +3,4 @@ matplotlib
 pandas
 scipy
 coverage
-python-coveralls
 codeclimate-test-reporter
