### 2017-05-05 a7ec7dc
Tweaked badge formatting
