### 2017-05-05 2264e95
Typographic tweaks in README
