### 2017-05-05 71c6461
Fixed typographical typo in README
