### 2017-05-05 48d40b4
More detailed testing info added to README
