### 2017-05-05 f4b2d2f
Small refactoring of simulation.run() method
