### 2017-05-05 24554f9
Simplified checks for whether a simulation has been initialised
