### 2017-05-06 56ea4c2
Brought requirements in setup.py up to date
