+++ setup.py
@@ -27,7 +27,8 @@ config = {
                           'matplotlib', 
                           'pandas', 
                           'scipy', 
-                          'python-coveralls' ],
+                          'coverage',
+                          'codeclimate-test-reporter' ],
     'license': 'MIT',
     'packages': ['lattice_mc'],
     'scripts': [],
