### 2017-05-10 7baec74
Require previous coverage version until the codeclimate reporter is updated
