+++ requirements.txt
@@ -2,5 +2,5 @@ numpy
 matplotlib
 pandas
 scipy
-coverage
+coverage==4.3.4
 codeclimate-test-reporter
