### 2017-05-22 #247 @bjmorgan
@arfon Would it be appropriate to bump the version number to e.g. 1.0.1 to differentiate from the as submitted version?
