+++ setup.py
@@ -12,7 +12,7 @@ except ImportError:
 def read(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
 
-VERSION = '1.0.0'
+VERSION = '1.0.1'
 
 config = {
     'description': 'A lattice-gas Monte-Carlo simulation tool',
