### 2017-05-23 039e58a
Updated README to include pip installation
