### 2017-05-23 e73a6ab
Added PyPI version badge to README
