### 2017-05-23 #247 @bjmorgan
The doi for the "as accepted" version (1.0.1) is 10.5281/zenodo.582402
The doi for the original submitted version (1.0.0) is 10.5281/zenodo.546262
