+++ setup.py
@@ -5,18 +5,17 @@ try:
 except ImportError:
     from distutils.core import setup
 
-# Utility function to read the README file.
-# Used for the long_description.  It's nice, because now 1) we have a top level
-# README file and 2) it's easier to type in the README file than to put a raw
-# string in below ...
-def read(fname):
-    return open(os.path.join(os.path.dirname(__file__), fname)).read()
+try:
+    import pypandoc
+    long_description = pypandoc.convert('README.md', 'rst')
+except ImportError:
+    long_description = open('README.md').read()
 
-VERSION = '1.0.1'
+VERSION = '1.0.2'
 
 config = {
     'description': 'A lattice-gas Monte-Carlo simulation tool',
-    'long_description': read('README.md'),
+    'long_description': long_description,
     'author': 'Benjamin J. Morgan',
     'author_email': 'b.j.morgan@bath.ac.uk',
     'url': 'https://github.com/bjmorgan/lattice_mc',
