### 2017-05-23 07f2f89
Added __version__ metadata to module
