### 2017-05-23 d1887f4
Added version metadata to module
