+++ lattice_mc/__init__.py
@@ -1,2 +1,4 @@
 from lattice_mc.simulation import Simulation
 from lattice_mc.options import Options
+
+__version__ = '1.0.3'
