### 2017-05-23 #247 @labarba
@asmit3 : Thank you for your review, and I hope you enjoyed being part of this new publishing model and you'll consider submitting to JOSS in the future!
