### 2017-05-23 518c398
removed duplicated author_email
