+++ setup.py
@@ -21,7 +21,6 @@ config = {
     'author_email': 'b.j.morgan@bath.ac.uk',
     'url': 'https://github.com/bjmorgan/lattice_mc',
     'download_url': "https://github.com/bjmorgan/lattice_mc/archive/%s.tar.gz" % (VERSION),
-    'author_email': 'b.j.morgan@bath.ac.uk',
     'version': VERSION,
     'install_requires': [ 'numpy', 
                           'matplotlib', 
