### 2017-05-26 #247 @arfon
Hi @bjmorgan - could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)). They are automatically compiled into the PDF when we process the paper.
