### 2017-05-26 3ec3074
Fixed citations
Fixed the citation to Murch Sol. Stat. Ionics 1982 in README.md
Converted the citation syntax in paper.md to cite references directly
cf. http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html
