### 2017-05-26 4fc9d03
bibtex strings fixed in paper.md
