+++ paper.bib
@@ -1,4 +1,4 @@
-@unpublished{Morgan_LLZO_LGMC,
+@unpublished{Morgan_LLZO,
    	Author = {Morgan, B. J.},
 	Note = {In Preparation}}	
 
@@ -59,7 +59,7 @@
 	Volume = {36},
 	Year = {1977}}
 
-@inbook{Voter_RadiationEffectsInSolids2007,
+@inbook{Voter_kMCmethod,
 	Address = {Dordrecht},
 	Author = {Voter, Arthur F.},
 	Booktitle = {Radiation Effects in Solids},
@@ -75,7 +75,7 @@
 	Year = {2007},
 	Bdsk-Url-1 = {http://dx.doi.org/10.1007/978-1-4020-5295-8_1}}
 
-@article{Catlow_SolidStateIonics1983,
+@article{Catlow_SolStatIonics1983,
 	Author = {Catlow, C. Richard A.},
 	Date-Added = {2012-12-20 17:14:00 +0000},
 	Date-Modified = {2012-12-20 17:14:37 +0000},
@@ -96,7 +96,7 @@
 	Volume = {27},
 	Year = {1964}}
 
-@inbook{Stoneham_IonicSolidsAtHighTemperatures,
+@inbook{Stoneham_IonicSolids1989,
 	Author = {Harding, J. F.},
         Title = {Defects and Transport in Ionic Solids},
         Date-Added = {2016-12-09 16:41:41 +0000},
