### 2017-05-26 #247 @bjmorgan
@arfon: I have updated the `paper.md`.

https://github.com/bjmorgan/lattice_mc/commit/3ec30748f24614b5404da739b64171d0549806b7
https://github.com/bjmorgan/lattice_mc/commit/4fc9d0380177dc1ca2994a2440acd11840be4d20
